PoseFit_Trainer/AI_PersonTrainer backend/requirements.txt
fastapi==0.110.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.6.3
email-validator==2.1.0
uvicorn==0.27.1
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.24.3
mediapipe==0.10.5
matplotlib==3.7.2
pandas==2.0.3
Pillow==10.0.0
jinja2==3.1.2
bcrypt==4.3.0
sounddevice==0.5.1
cryptography==44.0.2
requests==2.31.0
